171 different difficulty levels (#173)
* game mode determines the max number of enemies created

* select and unselect status of each game mode --- contain bugs

* able to save and load game state

* commit all

* game balance

* clean up exe files

* clean up

* fix button positions

* add assert()

* add assert()

* decrease health values of the enemies & Friend_boss_difficulty in easy mode

* clean up

* commit all

* clean up

* revert changes

+++ src/sub_systems/menu_system.cpp
@@ -29,8 +29,10 @@ MENU_OPTION MenuSystem::poll_start_menu() {
         }
     }
 
-    // destroy all menu elements after a choice is made by the user
-    if (selected_option != MENU_OPTION::NONE) {
+    
+    if (selected_option != MENU_OPTION::NONE 
+        && selected_option != MENU_OPTION::REGULAR_GAME_MODE 
+        && selected_option != MENU_OPTION::EASY_GAME_MODE) {// destroy all menu elements after a choice is made by the user
         destroy_menu();
     }
 
@@ -143,6 +145,8 @@ void MenuSystem::draw_start_menu() {
     create_menu_button({0.f, 50.f}, TEXTURE_ASSET_ID::MENU_START, MENU_OPTION::START_GAME);
     create_menu_button({0.f, -150.f}, TEXTURE_ASSET_ID::MENU_LOAD, MENU_OPTION::LOAD_GAME);
     create_menu_button({0.f, -350.f}, TEXTURE_ASSET_ID::MENU_EXIT, MENU_OPTION::EXIT_GAME);
+    create_menu_button({ -550.f, -150.f }, TEXTURE_ASSET_ID::MENU_EASY_MODE_UNSELECTED, MENU_OPTION::EASY_GAME_MODE);
+    create_menu_button({ -550.f, -350.f }, TEXTURE_ASSET_ID::MENU_REGULAR_MODE_UNSELECTED, MENU_OPTION::REGULAR_GAME_MODE);
 }
 
 void MenuSystem::draw_pause_menu() {
